# server/route/api/v1/swagger.yaml
basePath: /
definitions:
...
  api_v1.CustomizedProfile:
    properties:
      appearance:
        description: Appearance is the server default appearance.
        type: string
      description:
        description: Description is the server description.
        type: string
      faviconUrl:
        description: FaviconURL is the browser tab icon URL or data URL.
        type: string
      locale:
        description: Locale is the server default locale.
        type: string
      logoUrl:
        description: LogoURL is the url of logo image.
        type: string
      name:
        description: Name is the server name, default is `memos`
        type: string
    type: object
...
  api_v1.Role:
    enum:
    - HOST
    - ADMIN
    - USER
    type: string
    x-enum-varnames:
    - RoleHost
    - RoleAdmin
    - RoleUser
  api_v1.RowStatus:
    enum:
    - NORMAL
    - ARCHIVED
    type: string
    x-enum-varnames:
    - Normal
    - Archived
...
  api_v1.SystemSetting:
    properties:
      description:
        type: string
      name:
        $ref: '#/definitions/api_v1.SystemSettingName'
      value:
        description: Value is a JSON string with basic value.
        type: string
    type: object
  api_v1.SystemSettingName:
    enum:
    - server-id
    - secret-session
    - disable-public-memos
    - max-upload-size-mib
    - customized-profile
    - storage-service-id
    - local-storage-path
    - telegram-bot-token
    - memo-display-with-updated-ts
    type: string
    x-enum-varnames:
    - SystemSettingServerIDName
    - SystemSettingSecretSessionName
    - SystemSettingDisablePublicMemosName
    - SystemSettingMaxUploadSizeMiBName
    - SystemSettingCustomizedProfileName
    - SystemSettingStorageServiceIDName
    - SystemSettingLocalStoragePathName
    - SystemSettingTelegramBotTokenName
    - SystemSettingMemoDisplayWithUpdatedTsName
  api_v1.SystemStatus:
    properties:
      additionalScript:
        description: Additional script.
        type: string
      additionalStyle:
        description: Additional style.
        type: string
      allowSignUp:
        description: |-
          System settings
          Allow sign up.
        type: boolean
      customizedProfile:
        allOf:
        - $ref: '#/definitions/api_v1.CustomizedProfile'
        description: Customized server profile, including server name and external
          url.
      dbSize:
        type: integer
      disablePasswordLogin:
        description: Disable password login.
        type: boolean
      disablePublicMemos:
        description: Disable public memos.
        type: boolean
      host:
        $ref: '#/definitions/api_v1.User'
      localStoragePath:
        description: Local storage path.
        type: string
      maxUploadSizeMiB:
        description: Max upload size.
        type: integer
      memoDisplayWithUpdatedTs:
        description: Memo display with updated timestamp.
        type: boolean
      profile:
        $ref: '#/definitions/profile.Profile'
      storageServiceId:
        description: Storage service ID.
        type: integer
    type: object
...
  api_v1.UpsertSystemSettingRequest:
    properties:
      description:
        type: string
      name:
        $ref: '#/definitions/api_v1.SystemSettingName'
      value:
        type: string
    type: object
...
  api_v1.User:
    properties:
      avatarUrl:
        type: string
      createdTs:
        type: integer
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/api_v1.Role'
      rowStatus:
        allOf:
        - $ref: '#/definitions/api_v1.RowStatus'
        description: Standard fields
      updatedTs:
        type: integer
      username:
        description: Domain specific fields
        type: string
    type: object
...
  profile.Profile:
    properties:
      mode:
        description: Mode can be "prod" or "dev" or "demo"
        type: string
      version:
        description: Version is the current version of server
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/orgs/usememos/discussions
  description: A privacy-first, lightweight note-taking service.
  license:
    name: MIT License
    url: https://github.com/usememos/memos/blob/main/LICENSE
  title: memos API
  version: "1.0"
paths:
...
  /api/v1/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: System GetSystemStatus
          schema:
            $ref: '#/definitions/api_v1.SystemStatus'
        "401":
          description: Missing user in session | Unauthorized
        "500":
          description: Failed to find host user | Failed to find system setting list
            | Failed to unmarshal system setting customized profile value
      summary: Get system GetSystemStatus
      tags:
      - system
...
  /api/v1/system/setting:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: System setting list
          schema:
            items:
              $ref: '#/definitions/api_v1.SystemSetting'
            type: array
        "401":
          description: Missing user in session | Unauthorized
        "500":
          description: Failed to find user | Failed to find system setting list
      summary: Get a list of system settings
      tags:
      - system-setting
    post:
      consumes:
      - application/json
      parameters:
      - description: Request object.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_v1.UpsertSystemSettingRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Malformatted post system setting request | invalid system setting
        "401":
          description: Missing user in session | Unauthorized
        "403":
          description: Cannot disable passwords if no SSO identity provider is configured.
        "500":
          description: Failed to find user | Failed to upsert system setting
      summary: Create system setting
      tags:
      - system-setting
  /api/v1/system/vacuum:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: Database vacuumed
          schema:
            type: boolean
        "401":
          description: Missing user in session | Unauthorized
        "500":
          description: Failed to find user | Failed to ExecVacuum database
      summary: Vacuum the database
      tags:
      - system
...
swagger: "2.0"